Improve performance by caching repeated file glob results.
When processing failed test results, this action spent a decent amount
of time in repeated `resolvePath()` calls for the same paths if there
were multiple test failures for the same file path.

This introduces a cache so that repeated calls to `resolvePath()` don't
need to perform the globbing if the same file input is given repeatedly
across multiple tests.